Validate SessionInfo in command responses (#7)
Per the spec (https://github.com/teslamotors/vehicle-command/blob/main/pkg/protocol/protocol.md#recovering-from-synchronization-errors)

SessionInfo should be invalidated on an error if:
* The UUID doesn't match
* The HMAC tag does not match
* The clock time is out-of-order

This change removes the SessionInfo from the repository if any of these are true when a signing fault occurs, and performs a new handshake.

Additionally, if the session is valid, the session data is now updated from the responses. The counter is only updated if the epochs did not match.

+++ src/main/kotlin/com/boltfortesla/teslafleetsdk/handshake/HandshakerImpl.kt
@@ -21,7 +21,6 @@ import com.tesla.generated.universalmessage.destination
 import com.tesla.generated.universalmessage.routableMessage
 import com.tesla.generated.universalmessage.sessionInfoRequest
 import java.util.Base64
-import java.util.concurrent.atomic.AtomicInteger
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 
@@ -100,7 +99,7 @@ internal class HandshakerImpl(
       return SessionInfo(
         sessionInfo.epoch.toByteArray(),
         sessionInfo.clockTime,
-        AtomicInteger(sessionInfo.counter),
+        sessionInfo.counter,
         sharedSecret,
       )
     }
