+++ src/test/kotlin/com/boltfortesla/teslafleetsdk/handshake/SessionInfoTest.kt
@@ -1,7 +1,6 @@
 package com.boltfortesla.teslafleetsdk.handshake
 
 import com.google.common.truth.Truth.assertThat
-import java.util.concurrent.atomic.AtomicInteger
 import org.junit.Test
 
 class SessionInfoTest {
@@ -9,16 +8,11 @@ class SessionInfoTest {
   @Test
   fun equals() {
     val sessionInfo =
-      SessionInfo(
-        "epoch".toByteArray(),
-        clockTime = 0,
-        counter = AtomicInteger(0),
-        "sharedSecret".toByteArray()
-      )
+      SessionInfo("epoch".toByteArray(), clockTime = 0, counter = 0, "sharedSecret".toByteArray())
 
     assertThat(sessionInfo).isNotEqualTo(sessionInfo.copy(epoch = "other".toByteArray()))
     assertThat(sessionInfo).isNotEqualTo(sessionInfo.copy(clockTime = 1))
-    assertThat(sessionInfo).isNotEqualTo(sessionInfo.copy(counter = AtomicInteger(1)))
+    assertThat(sessionInfo).isNotEqualTo(sessionInfo.copy(counter = 1))
     assertThat(sessionInfo).isNotEqualTo(sessionInfo.copy(sharedSecret = "other".toByteArray()))
   }
 }
